refactor(codegen): reorganize codegen/ file structure
- Rename wasm_builder.rs → component_context.rs
  (contains only ComponentModelContext; old name was misleading)
- Rename wasm_postprocess.rs → postprocess.rs
  (drop redundant wasm_ prefix inside the codegen module)
- Merge component.rs (18-line thin wrapper) into component_gen.rs,
  then rename component_gen.rs → component.rs
  (eliminates artificial two-file split; removes unused _wir parameter)
- Update codegen.rs module declarations and build_component call site
